refactor: general_token
