+++ src/features/auth/domain/usecases/dto.rs
@@ -1,7 +1,6 @@
 use chrono::NaiveDateTime;
 use diesel::Insertable;
 use uuid::Uuid;
-use validator::Validate;
 
 use crate::{camel_case_struct, schema::login_history};
 
@@ -14,29 +13,6 @@ pub struct LoginHistoryParams {
     pub device_info: String,
     pub os_info: String,
     pub fcm_token: String,
-}
-
-
-
-camel_case_struct!(GeneralTokenParams {
-  #[validate(
-    required(message = "field is required"),
-    length(min = 1, message = "Can't be empty"),
-  )]
-   client_id: Option<String>,
-  #[validate(
-    required(message = "field is required"),
-    length(min = 1, message = "Can't be empty"),
-  )]
-   client_secret: Option<String>
-});
-
-impl GeneralTokenParams {
-    pub fn verify(&self) -> bool {
-        self.validate().is_ok()
-            && self.client_id.as_deref() == Some(env!("CLIENT_ID"))
-            && self.client_secret.as_deref() == Some(env!("CLIENT_SECRET"))
-    }
 }
 
 camel_case_struct!(UpdatePasswordParams {
