+++ src/features/auth/domain/usecases/interface.rs
@@ -5,7 +5,7 @@ use crate::{
     core::types::AppResult,
     features::auth::{
         data::models::{auth_token::AuthToken, login_history::LoginHistory},
-        domain::{entity::auth_response::AuthResponse, usecases::dto::*},
+        domain::usecases::dto::*,
     },
 };
 
@@ -14,6 +14,6 @@ pub trait IAuthService: Send + Sync {
     // fn logout(&self, user: Uuid, login_session: Uuid) -> AppResult<()>;
     fn verify_token(&self, params: &TokenData<AuthToken>) -> AppResult<Uuid>;
     fn login_session(&self, user: Uuid) -> AppResult<Vec<LoginHistory>>;
-    fn general_token(&self, token: GeneralTokenParams) -> AppResult<AuthResponse>;
+    // fn general_token(&self, token: GeneralTokenParams) -> AppResult<AuthResponse>;
     fn update_password(&self, user: Uuid, params: UpdatePasswordParams) -> AppResult<()>;
 }
