+++ src/features/auth/domain/usecases/mod.rs
@@ -3,3 +3,4 @@ pub mod interface;
 pub mod service;
 pub mod login;
 pub mod logout;
+pub mod general_token;
